requirements/base.txt
Django==5.0.3
django-environ-0.11.2
djangorestframework-3.14.0
drf-spectacular-0.27.1
django_debug_toolbar==3.8.1
djangorestframework-simplejwt-5.3.1
